requirements.txt
python_pptx
pandas
matplotlib
imageio


